Add Bamboo deploy tasks and crontabs
A copy of the Bamboo deploy tasks are being committed here for
safekeeping and to track the history of changes.

The cron jobs for each server are being committed here for
safekeeping.  Updating a crontab file in the repo will also
update the actual cron jobs on the server during the next deploy.